Bakc to working again. imagefile list is only initialised once and freed once inside main.c - no other memory allocation for game artwork parsing.

+++ src/timers.c
@@ -31,7 +31,7 @@ int timers_FireArt(clock_t last){
 	// Returns true if the timeout since the last input has
 	// exceeded that to fire the artwork display routine
 	
-	int t;
+	long long int t;
 	
 	t = clock() - last;
